refactor: extract all remaining admin endpoints from orchestrator (#550)
Phase 4.5 of orchestrator decomposition (Strangler Fig):

Extract 17 unique endpoints into 3 new routers:
- modules/speech/router_voices.py: voice selection + test (4 endpoints)
- modules/llm/router_models.py: HuggingFace model management (10 endpoints)
- modules/monitoring/router_logs.py: log viewing + SSE streaming (3 endpoints)

Remove 35 dead duplicate endpoints that were shadowed by module routers:
- 25 LLM endpoints (duplicated in modules/llm/router.py)
- 6 monitoring endpoints (duplicated in modules/monitoring/router_monitor.py)
- 6 services endpoints (duplicated in modules/speech/router_services.py)

Also removed: 18 dead Pydantic models, get_current_tts_service() helper,
and 16 unused imports (time, Dict, List, Depends, HTTPException, FileResponse,
StreamingResponse, BaseModel, User, get_current_user, require_permission,
PROVIDER_TYPES, async_audit_logger, get_model_manager, get_service_manager,
get_system_monitor).

Orchestrator: 2471 → 1121 lines (-1350, -55%)
Zero inline @app endpoints remain — only startup/shutdown + static files.
