+++ orchestrator.py
@@ -7,20 +7,16 @@ STT (Whisper) -> LLM (vLLM / Cloud) -> TTS (XTTS v2)
 import asyncio
 import logging
 import os
-import time
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import Optional
 
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException, Request
+from fastapi import FastAPI, Request
 from fastapi.responses import (
-    FileResponse,
     RedirectResponse,
     Response,
-    StreamingResponse,
 )
-from pydantic import BaseModel
 from slowapi import _rate_limit_exceeded_handler
 from slowapi.errors import RateLimitExceeded
 
@@ -62,16 +58,10 @@ from app.security_headers import (
     SECURITY_HEADERS_ENABLED,
     SecurityHeadersMiddleware,
 )
-from auth_manager import (
-    User,
-    get_current_user,
-    require_permission,
-)
 
 # Cloud LLM service for multi-provider support
-from cloud_llm_service import PROVIDER_TYPES, CloudLLMService
+from cloud_llm_service import CloudLLMService
 from db.integration import (
-    async_audit_logger,
     async_cloud_provider_manager,
     async_faq_manager,
     async_preset_manager,
@@ -82,8 +72,8 @@ from db.integration import (
     shutdown_database,
 )
 
+
 # Database integration
-from model_manager import get_model_manager
 
 
 # Multi-bot manager
@@ -95,8 +85,6 @@ except ImportError:
     PIPER_AVAILABLE = False
     PiperTTSService = None
 
-from service_manager import get_service_manager
-
 
 try:
     from stt_service import STTService
@@ -105,8 +93,6 @@ try:
 except ImportError:
     STT_AVAILABLE = False
     STTService = None
-
-from system_monitor import get_system_monitor
 
 
 # Импорты наших сервисов
@@ -228,6 +214,24 @@ if DEPLOYMENT_MODE != "cloud":
 
     app.include_router(llm_finetune_router)
     app.include_router(tts_finetune_router)
+    # Voice + model management (Phase 4.5) — GPU-only
+    from modules.llm.router_models import router as models_router
+    from modules.speech.router_voices import router as voices_router
+
+    app.include_router(voices_router)
+    app.include_router(models_router)
+
+# Logs router (available in all modes)
+from modules.monitoring.router_logs import router as logs_router  # noqa: E402
+
+
+app.include_router(logs_router)
+
+# Widget public endpoints
+from modules.channels.widget.router_public import router as widget_public_router  # noqa: E402
+
+
+app.include_router(widget_public_router)
 
 # Глобальные сервисы
 voice_service: Optional["VoiceCloneService"] = None  # XTTS (Марина) - GPU CC >= 7.0
@@ -1058,1361 +1062,7 @@ async def shutdown_event():
 # Vue 3 admin panel served from /admin (see Static Files section below)
 
 
-# ============== Admin API Endpoints ==============
-
-# ============== Admin Models ==============
-
-
-class AdminTTSPresetRequest(BaseModel):
-    """Запрос на изменение пресета TTS"""
-
-    preset: str  # warm, calm, energetic, natural, neutral
-
-
-class AdminLLMPromptRequest(BaseModel):
-    """Запрос на изменение системного промпта"""
-
-    prompt: str
-
-
-class AdminLLMModelRequest(BaseModel):
-    """Запрос на изменение модели LLM"""
-
-    model: str  # gemini-2.5-flash, gemini-2.5-pro
-
-
-class AdminTTSTestRequest(BaseModel):
-    """Запрос на тестовый синтез"""
-
-    text: str
-    preset: str = "natural"
-
-
-@app.get("/admin/llm/prompt")
-async def admin_get_llm_prompt():
-    """Получить текущий системный промпт LLM"""
-    if llm_service:
-        persona = getattr(llm_service, "current_persona", None) or os.getenv(
-            "SECRETARY_PERSONA", "anna"
-        )
-        return {
-            "prompt": llm_service.system_prompt,
-            "model": llm_service.model_name,
-            "persona": persona,
-        }
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.post("/admin/llm/prompt")
-async def admin_set_llm_prompt(request: AdminLLMPromptRequest):
-    """Установить новый системный промпт LLM"""
-    if llm_service:
-        llm_service.set_system_prompt(request.prompt)
-        return {
-            "status": "ok",
-            "prompt": request.prompt[:100] + "..." if len(request.prompt) > 100 else request.prompt,
-        }
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.get("/admin/llm/model")
-async def admin_get_llm_model():
-    """Получить текущую модель LLM"""
-    if llm_service:
-        return {"model": llm_service.model_name}
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.post("/admin/llm/model")
-async def admin_set_llm_model(request: AdminLLMModelRequest):
-    """Изменить модель LLM"""
-    allowed_models = ["gemini-2.5-flash", "gemini-2.5-pro", "gemini-2.0-flash"]
-
-    if request.model not in allowed_models:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Неизвестная модель: {request.model}. Доступные: {allowed_models}",
-        )
-
-    if llm_service:
-        try:
-            llm_service.set_model(request.model)
-            return {"status": "ok", "model": request.model}
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=str(e))
-
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.delete("/admin/llm/history")
-async def admin_clear_llm_history():
-    """Очистить историю диалога LLM"""
-    if llm_service:
-        count = len(llm_service.conversation_history)
-        llm_service.reset_conversation()
-        return {"status": "ok", "cleared_messages": count}
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.get("/admin/llm/history")
-async def admin_get_llm_history():
-    """Получить историю диалога LLM"""
-    if llm_service:
-        return {
-            "history": llm_service.conversation_history,
-            "count": len(llm_service.conversation_history),
-        }
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-# ============== Voice Selection API ==============
-
-
-class AdminVoiceRequest(BaseModel):
-    voice: str  # anna / marina / dmitri / irina
-
-
-@app.get("/admin/voices")
-async def admin_get_voices():
-    """Получить список всех доступных голосов"""
-    voices = []
-
-    # XTTS голос (Анна) - требует GPU CC >= 7.0 (по умолчанию)
-    if anna_voice_service:
-        voices.append(
-            {
-                "id": "anna",
-                "name": "Анна (XTTS)",
-                "engine": "xtts",
-                "description": "Клонированный голос Гули (XTTS v2, GPU CC >= 7.0)",
-                "available": True,
-                "samples_count": len(anna_voice_service.voice_samples),
-                "default": True,
-            }
-        )
-
-    # XTTS голос (Марина) - требует GPU CC >= 7.0
-    if voice_service:
-        voices.append(
-            {
-                "id": "marina",
-                "name": "Марина (XTTS)",
-                "engine": "xtts",
-                "description": "Клонированный голос Лидии (XTTS v2, GPU CC >= 7.0)",
-                "available": True,
-                "samples_count": len(voice_service.voice_samples),
-            }
-        )
-
-    # OpenVoice голос (Марина) - работает на GPU CC 6.1+
-    if openvoice_service:
-        voices.append(
-            {
-                "id": "marina_openvoice",
-                "name": "Марина (OpenVoice)",
-                "engine": "openvoice",
-                "description": "Клонированный голос (OpenVoice v2, GPU CC 6.1+)",
-                "available": True,
-                "samples_count": len(openvoice_service.voice_samples)
-                if openvoice_service.voice_samples
-                else 0,
-            }
-        )
-
-    # Piper голоса (CPU)
-    if piper_service:
-        piper_voices = piper_service.get_available_voices()
-        for voice_id, info in piper_voices.items():
-            voices.append(
-                {
-                    "id": voice_id,
-                    "name": info["name"],
-                    "engine": "piper",
-                    "description": info["description"],
-                    "available": info["available"],
-                }
-            )
-
-    return {
-        "voices": voices,
-        "current": current_voice_config,
-    }
-
-
-@app.get("/admin/voice")
-async def admin_get_current_voice():
-    """Получить текущий выбранный голос"""
-    return current_voice_config
-
-
-@app.post("/admin/voice")
-async def admin_set_voice(request: AdminVoiceRequest):
-    """Установить активный голос"""
-    global current_voice_config
-
-    voice_id = request.voice.lower()
-
-    # Проверяем доступность голоса
-    if voice_id == "anna":
-        if not anna_voice_service:
-            raise HTTPException(
-                status_code=503, detail="XTTS service (Анна) not available (requires GPU CC >= 7.0)"
-            )
-        current_voice_config = {"engine": "xtts", "voice": "anna"}
-        logger.info("🎤 Голос изменён на: Анна (XTTS)")
-
-    elif voice_id == "marina":
-        if not voice_service:
-            raise HTTPException(
-                status_code=503,
-                detail="XTTS service (Марина) not available (requires GPU CC >= 7.0)",
-            )
-        current_voice_config = {"engine": "xtts", "voice": "marina"}
-        logger.info("🎤 Голос изменён на: Марина (XTTS)")
-
-    elif voice_id == "marina_openvoice":
-        if not openvoice_service:
-            raise HTTPException(status_code=503, detail="OpenVoice service not available")
-        current_voice_config = {"engine": "openvoice", "voice": "marina_openvoice"}
-        logger.info("🎤 Голос изменён на: Марина (OpenVoice)")
-
-    elif voice_id in ["dmitri", "irina"]:
-        if not piper_service:
-            raise HTTPException(status_code=503, detail="Piper TTS service not available")
-        piper_voices = piper_service.get_available_voices()
-        if voice_id not in piper_voices or not piper_voices[voice_id]["available"]:
-            raise HTTPException(status_code=400, detail=f"Voice model not found: {voice_id}")
-        current_voice_config = {"engine": "piper", "voice": voice_id}
-        logger.info(f"🎤 Голос изменён на: {piper_voices[voice_id]['name']} (Piper)")
-
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Unknown voice: {voice_id}. Available: anna, marina, marina_openvoice, dmitri, irina",
-        )
-
-    # Sync with service container for modular routers
-    from app.dependencies import get_container
-
-    container = get_container()
-    container.current_voice_config = current_voice_config
-
-    return {"status": "ok", **current_voice_config}
-
-
-@app.post("/admin/voice/test")
-async def admin_test_voice(request: AdminVoiceRequest):
-    """Тестовый синтез выбранным голосом"""
-    voice_id = request.voice.lower()
-    test_text = "Здравствуйте! Это тестовое сообщение для проверки голоса."
-
-    output_path = TEMP_DIR / f"voice_test_{voice_id}_{int(time.time())}.wav"
-
-    try:
-        if voice_id == "anna":
-            if not anna_voice_service:
-                raise HTTPException(
-                    status_code=503, detail="XTTS (Анна) not available (requires GPU CC >= 7.0)"
-                )
-            anna_voice_service.synthesize_to_file(test_text, str(output_path), preset="natural")
-
-        elif voice_id == "marina":
-            if not voice_service:
-                raise HTTPException(
-                    status_code=503, detail="XTTS (Марина) not available (requires GPU CC >= 7.0)"
-                )
-            voice_service.synthesize_to_file(test_text, str(output_path), preset="natural")
-
-        elif voice_id == "marina_openvoice":
-            if not openvoice_service:
-                raise HTTPException(status_code=503, detail="OpenVoice not available")
-            openvoice_service.synthesize_to_file(test_text, str(output_path), language="ru")
-
-        elif voice_id in ["dmitri", "irina"]:
-            if not piper_service:
-                raise HTTPException(status_code=503, detail="Piper not available")
-            piper_service.synthesize_to_file(test_text, str(output_path), voice=voice_id)
-
-        else:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Unknown voice: {voice_id}. Available: anna, marina, marina_openvoice, dmitri, irina",
-            )
-
-        return FileResponse(output_path, media_type="audio/wav", filename=f"test_{voice_id}.wav")
-
-    except Exception as e:
-        logger.error(f"❌ Ошибка тестового синтеза: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-def get_current_tts_service():
-    """Возвращает текущий TTS сервис и параметры на основе конфигурации"""
-    engine = current_voice_config["engine"]
-    voice = current_voice_config["voice"]
-
-    if engine == "xtts" and voice == "anna":
-        return anna_voice_service, {"preset": "natural"}
-    elif engine == "xtts" and voice == "marina":
-        return voice_service, {"preset": "natural"}
-    elif engine == "piper":
-        return piper_service, {"voice": voice}
-    else:
-        # Default to anna if available
-        return anna_voice_service or voice_service, {"preset": "natural"}
-
-
-# ============== Extended Admin API Endpoints ==============
-
-
-# Pydantic models for new endpoints
-class AdminBackendRequest(BaseModel):
-    backend: str  # "vllm" or "cloud:{provider_id}"
-    stop_unused: bool = False  # Остановить неиспользуемый сервис (vLLM) для освобождения GPU
-
-
-class CloudProviderCreate(BaseModel):
-    """Create cloud LLM provider"""
-
-    name: str
-    provider_type: str  # gemini, kimi, openai, claude, deepseek, custom
-    api_key: Optional[str] = None
-    base_url: Optional[str] = None
-    model_name: str = ""
-    enabled: bool = True
-    is_default: bool = False
-    config: Optional[Dict] = None
-    description: Optional[str] = None
-
-
-class CloudProviderUpdate(BaseModel):
-    """Update cloud LLM provider"""
-
-    name: Optional[str] = None
-    provider_type: Optional[str] = None
-    api_key: Optional[str] = None
-    base_url: Optional[str] = None
-    model_name: Optional[str] = None
-    enabled: Optional[bool] = None
-    is_default: Optional[bool] = None
-    config: Optional[Dict] = None
-    description: Optional[str] = None
-
-
-class AdminPersonaRequest(BaseModel):
-    persona: str  # "anna" or "marina"
-
-
-class AdminLLMParamsRequest(BaseModel):
-    temperature: Optional[float] = None
-    max_tokens: Optional[int] = None
-    top_p: Optional[float] = None
-    repetition_penalty: Optional[float] = None
-
-
-class AdminXTTSParamsRequest(BaseModel):
-    temperature: Optional[float] = None
-    repetition_penalty: Optional[float] = None
-    top_k: Optional[int] = None
-    top_p: Optional[float] = None
-    speed: Optional[float] = None
-    gpt_cond_len: Optional[int] = None
-    gpt_cond_chunk_len: Optional[int] = None
-
-
-class AdminPiperParamsRequest(BaseModel):
-    speed: float = 1.0
-
-
-class AdminCustomPresetRequest(BaseModel):
-    name: str
-    params: dict
-
-
-class AdminFAQRequest(BaseModel):
-    trigger: str
-    response: str
-
-
-class AdminFAQTestRequest(BaseModel):
-    text: str
-
-
-class AdminWidgetConfigRequest(BaseModel):
-    enabled: bool = True
-    title: str = "AI Ассистент"
-    greeting: str = "Здравствуйте! Компания Шаервэй Ди-Иджитал, чем могу помочь?"
-    placeholder: str = "Введите сообщение..."
-    primary_color: str = "#c2410c"
-    position: str = "right"  # "left" or "right"
-    allowed_domains: List[str] = []
-    tunnel_url: str = ""
-
-
-class AdminTelegramConfigRequest(BaseModel):
-    enabled: bool = False
-    bot_token: str = ""
-    api_url: str = "http://localhost:8002"
-    allowed_users: List[int] = []
-    admin_users: List[int] = []
-    welcome_message: str = "Здравствуйте! Я AI-ассистент компании Шаервэй. Чем могу помочь?"
-    unauthorized_message: str = "Извините, у вас нет доступа к этому боту."
-    error_message: str = "Произошла ошибка. Попробуйте позже."
-    typing_enabled: bool = True
-
-
-class AdminAuditQueryRequest(BaseModel):
-    action: Optional[str] = None
-    resource: Optional[str] = None
-    user_id: Optional[str] = None
-    from_date: Optional[str] = None  # ISO format
-    to_date: Optional[str] = None  # ISO format
-    limit: int = 100
-    offset: int = 0
-
-
-# ============== Chat Models & Manager ==============
-
-
-class ChatMessageModel(BaseModel):
-    id: str
-    role: str
-    content: str
-    timestamp: str
-    edited: bool = False
-
-
-class ChatSessionModel(BaseModel):
-    id: str
-    title: str
-    messages: List[ChatMessageModel]
-    system_prompt: Optional[str] = None
-    created: str
-    updated: str
-
-
-class CreateSessionRequest(BaseModel):
-    title: Optional[str] = None
-    system_prompt: Optional[str] = None
-
-
-class UpdateSessionRequest(BaseModel):
-    title: Optional[str] = None
-    system_prompt: Optional[str] = None
-
-
-class LLMOverrideConfig(BaseModel):
-    llm_backend: Optional[str] = None  # "vllm" or "cloud:provider-id"
-    system_prompt: Optional[str] = None
-    llm_params: Optional[dict] = None
-
-
-class SendMessageRequest(BaseModel):
-    content: str
-    llm_override: Optional[LLMOverrideConfig] = None
-
-
-class EditMessageRequest(BaseModel):
-    content: str
-
-
-# ============== Services Endpoints ==============
-
-
-@app.get("/admin/services/status")
-async def admin_services_status():
-    """Получить статус всех сервисов"""
-    manager = get_service_manager()
-    status = manager.get_all_status()
-
-    # Добавляем статус внутренних сервисов из orchestrator
-    status["services"]["xtts_anna"]["is_running"] = anna_voice_service is not None
-    status["services"]["xtts_marina"]["is_running"] = voice_service is not None
-    status["services"]["piper"]["is_running"] = piper_service is not None
-    status["services"]["openvoice"]["is_running"] = openvoice_service is not None
-    status["services"]["orchestrator"]["is_running"] = True
-
-    return status
-
-
-@app.post("/admin/services/{service}/start")
-async def admin_start_service(service: str):
-    """Запустить сервис"""
-    manager = get_service_manager()
-    return await manager.start_service(service)
-
-
-@app.post("/admin/services/{service}/stop")
-async def admin_stop_service(service: str):
-    """Остановить сервис"""
-    manager = get_service_manager()
-    return await manager.stop_service(service)
-
-
-@app.post("/admin/services/{service}/restart")
-async def admin_restart_service(service: str):
-    """Перезапустить сервис"""
-    manager = get_service_manager()
-    return await manager.restart_service(service)
-
-
-@app.post("/admin/services/start-all")
-async def admin_start_all_services():
-    """Запустить все внешние сервисы"""
-    manager = get_service_manager()
-    results = {}
-    for service in ["vllm"]:  # Только внешние сервисы
-        results[service] = await manager.start_service(service)
-    return {"status": "ok", "results": results}
-
-
-@app.post("/admin/services/stop-all")
-async def admin_stop_all_services():
-    """Остановить все внешние сервисы"""
-    manager = get_service_manager()
-    results = {}
-    for service in ["vllm"]:
-        results[service] = await manager.stop_service(service)
-    return {"status": "ok", "results": results}
-
-
-# ============== Logs Endpoints ==============
-
-
-@app.get("/admin/logs")
-async def admin_list_logs():
-    """Список доступных логов"""
-    manager = get_service_manager()
-    return {"logs": manager.get_available_logs()}
-
-
-@app.get("/admin/logs/{logfile}")
-async def admin_read_log(
-    logfile: str, lines: int = 100, offset: int = 0, search: Optional[str] = None
-):
-    """Прочитать лог файл"""
-    manager = get_service_manager()
-    return manager.read_log(logfile, lines=lines, offset=offset, search=search)
-
-
-@app.get("/admin/logs/stream/{logfile}")
-async def admin_stream_log(
-    logfile: str,
-    user: User = Depends(require_permission("system", "view")),
-):
-    """SSE streaming логов"""
-    manager = get_service_manager()
-
-    async def generate():
-        async for data in manager.stream_log(logfile):
-            yield f"data: {data}\n\n"
-
-    return StreamingResponse(
-        generate(),
-        media_type="text/event-stream",
-        headers={"Cache-Control": "no-cache", "Connection": "keep-alive"},
-    )
-
-
-# ============== LLM Enhanced Endpoints ==============
-
-
-@app.get("/admin/llm/backend")
-async def admin_get_llm_backend():
-    """Получить текущий LLM backend"""
-    if llm_service:
-        # Detect backend type
-        if isinstance(llm_service, CloudLLMService):
-            backend = f"cloud:{llm_service.provider_id}"
-        elif hasattr(llm_service, "api_url"):
-            backend = "vllm"
-        else:
-            backend = "unknown"
-
-        return {
-            "backend": backend,
-            "model": getattr(llm_service, "model_name", "unknown"),
-            "api_url": getattr(llm_service, "api_url", None),
-            "provider_type": getattr(llm_service, "provider_type", None),
-        }
-    return {"backend": "none", "error": "LLM service not initialized"}
-
-
-@app.get("/admin/llm/models")
-async def admin_get_llm_models():
-    """
-    Получить список доступных моделей vLLM и текущую модель.
-    Возвращает информацию о Qwen, Llama, DeepSeek и других моделях.
-    """
-    from vllm_llm_service import AVAILABLE_MODELS
-
-    result = {
-        "available_models": AVAILABLE_MODELS,
-        "current_model": None,
-        "loaded_models": [],
-        "backend": "none",
-    }
-
-    if llm_service:
-        if isinstance(llm_service, CloudLLMService):
-            result["backend"] = f"cloud:{llm_service.provider_id}"
-            result["current_model"] = {
-                "id": llm_service.provider_id,
-                "name": getattr(llm_service, "model_name", "unknown"),
-                "description": f"Cloud: {getattr(llm_service, 'provider_type', 'unknown')}",
-                "available": True,
-            }
-        elif hasattr(llm_service, "api_url"):
-            result["backend"] = "vllm"
-            if hasattr(llm_service, "get_current_model_info"):
-                result["current_model"] = llm_service.get_current_model_info()
-                result["loaded_models"] = llm_service.get_loaded_models()
-
-    return result
-
-
-@app.post("/admin/llm/backend")
-async def admin_set_llm_backend(
-    request: AdminBackendRequest, user: User = Depends(get_current_user)
-):
-    """Переключить LLM backend с горячей перезагрузкой сервиса"""
-    global LLM_BACKEND, llm_service
-
-    # Auto-convert "gemini" to default cloud Gemini provider
-    if request.backend == "gemini":
-        gemini_provider = await _get_or_create_default_gemini_provider()
-        if gemini_provider:
-            return await _switch_to_cloud_provider(gemini_provider["id"], request.stop_unused, user)
-        raise HTTPException(
-            status_code=400,
-            detail="Cannot switch to Gemini: no GEMINI_API_KEY configured. "
-            "Create a Gemini cloud provider first.",
-        )
-
-    # Check if it's a cloud provider
-    if request.backend.startswith("cloud:"):
-        provider_id = request.backend.split(":", 1)[1]
-        return await _switch_to_cloud_provider(provider_id, request.stop_unused, user)
-
-    if request.backend != "vllm":
-        raise HTTPException(
-            status_code=400,
-            detail="Invalid backend. Use 'vllm' or 'cloud:{provider_id}'",
-        )
-
-    # Проверяем текущий бэкенд
-    current_backend = (
-        "vllm"
-        if (
-            llm_service
-            and hasattr(llm_service, "api_url")
-            and not isinstance(llm_service, CloudLLMService)
-        )
-        else "cloud"
-    )
-    if request.backend == current_backend:
-        return {
-            "status": "ok",
-            "backend": request.backend,
-            "message": f"Уже используется {request.backend}",
-        }
-
-    manager = get_service_manager()
-
-    try:
-        if request.backend == "vllm":
-            # Переключение на vLLM
-            logger.info("🔄 Переключение на vLLM...")
-
-            # Проверяем, запущен ли vLLM
-            vllm_status = manager.get_service_status("vllm")
-
-            if not vllm_status.get("is_running"):
-                # Запускаем vLLM
-                logger.info("🚀 Запуск vLLM...")
-                start_result = await manager.start_service("vllm")
-                if start_result.get("status") != "ok":
-                    raise HTTPException(
-                        status_code=503,
-                        detail=f"Не удалось запустить vLLM: {start_result.get('message', 'Unknown error')}",
-                    )
-
-                # Ждём готовности vLLM (до 120 секунд)
-                logger.info("⏳ Ожидание готовности vLLM...")
-                import httpx
-
-                # Нормализуем URL (удаляем trailing /v1)
-                vllm_url = os.getenv("VLLM_API_URL", "http://localhost:11434").rstrip("/")
-                if vllm_url.endswith("/v1"):
-                    vllm_url = vllm_url[:-3]
-
-                for i in range(60):  # 60 * 2 = 120 секунд
-                    await asyncio.sleep(2)
-                    try:
-                        async with httpx.AsyncClient() as client:
-                            resp = await client.get(f"{vllm_url}/v1/models", timeout=5.0)
-                            if resp.status_code == 200:
-                                logger.info(f"✅ vLLM готов (попытка {i + 1})")
-                                break
-                    except Exception:
-                        pass
-                else:
-                    raise HTTPException(
-                        status_code=503, detail=f"vLLM не стал доступен за 120 секунд ({vllm_url})"
-                    )
-
-            # Создаём новый vLLM сервис
-            if VLLMLLMService is None:
-                raise HTTPException(status_code=503, detail="VLLMLLMService не доступен")
-
-            new_service = VLLMLLMService()
-            if not new_service.is_available():
-                raise HTTPException(status_code=503, detail="vLLM запущен, но не отвечает на API")
-
-            llm_service = new_service
-            LLM_BACKEND = "vllm"
-            os.environ["LLM_BACKEND"] = "vllm"
-
-            logger.info("✅ Переключено на vLLM")
-
-            # Audit log
-            await async_audit_logger.log(
-                action="update",
-                resource="config",
-                resource_id="llm_backend",
-                user_id=user.username,
-                details={"backend": "vllm", "model": getattr(llm_service, "model_name", "unknown")},
-            )
-
-            return {
-                "status": "ok",
-                "backend": "vllm",
-                "model": getattr(llm_service, "model_name", "unknown"),
-                "message": "Переключено на vLLM",
-            }
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"❌ Ошибка переключения бэкенда: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-async def _switch_to_cloud_provider(provider_id: str, stop_unused: bool, user: User):
-    """Helper function to switch to a cloud provider"""
-    global LLM_BACKEND, llm_service
-
-    provider_config = await async_cloud_provider_manager.get_provider_with_key(provider_id)
-    if not provider_config:
-        raise HTTPException(status_code=404, detail=f"Provider {provider_id} not found")
-
-    if not provider_config.get("enabled"):
-        raise HTTPException(status_code=400, detail=f"Provider {provider_id} is disabled")
-
-    if not provider_config.get("api_key"):
-        raise HTTPException(
-            status_code=400, detail=f"Provider {provider_id} has no API key configured"
-        )
-
-    try:
-        new_service = CloudLLMService(provider_config)
-        if not new_service.is_available():
-            raise HTTPException(status_code=503, detail=f"Provider {provider_id} is not responding")
-
-        llm_service = new_service
-        LLM_BACKEND = f"cloud:{provider_id}"
-        os.environ["LLM_BACKEND"] = LLM_BACKEND
-
-        # Optionally stop vLLM to free GPU
-        if stop_unused:
-            manager = get_service_manager()
-            vllm_status = manager.get_service_status("vllm")
-            if vllm_status.get("is_running"):
-                logger.info("🛑 Stopping vLLM to free GPU memory...")
-                await manager.stop_service("vllm")
-
-        logger.info(f"✅ Switched to cloud provider: {provider_config.get('name')}")
-
-        # Audit log
-        await async_audit_logger.log(
-            action="update",
-            resource="config",
-            resource_id="llm_backend",
-            user_id=user.username,
-            details={
-                "backend": f"cloud:{provider_id}",
-                "provider_type": provider_config.get("provider_type"),
-                "model": provider_config.get("model_name"),
-            },
-        )
-
-        return {
-            "status": "ok",
-            "backend": f"cloud:{provider_id}",
-            "provider_id": provider_id,
-            "provider_type": provider_config.get("provider_type"),
-            "model": provider_config.get("model_name"),
-            "message": f"Switched to cloud provider: {provider_config.get('name')}",
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"❌ Error switching to cloud provider: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-# ============== Cloud LLM Providers API ==============
-
-
-@app.get("/admin/llm/providers")
-async def admin_list_cloud_providers(enabled_only: bool = False):
-    """List all cloud LLM providers"""
-    providers = await async_cloud_provider_manager.list_providers(enabled_only)
-    return {
-        "providers": providers,
-        "provider_types": PROVIDER_TYPES,
-    }
-
-
-@app.get("/admin/llm/providers/{provider_id}")
-async def admin_get_cloud_provider(
-    provider_id: str, include_key: bool = False, user: User = Depends(get_current_user)
-):
-    """Get cloud provider by ID"""
-    if include_key:
-        provider = await async_cloud_provider_manager.get_provider_with_key(provider_id)
-    else:
-        provider = await async_cloud_provider_manager.get_provider(provider_id)
-
-    if not provider:
-        raise HTTPException(status_code=404, detail="Provider not found")
-    return {"provider": provider}
-
-
-@app.post("/admin/llm/providers")
-async def admin_create_cloud_provider(
-    data: CloudProviderCreate, user: User = Depends(get_current_user)
-):
-    """Create new cloud LLM provider"""
-    try:
-        provider = await async_cloud_provider_manager.create_provider(
-            name=data.name,
-            provider_type=data.provider_type,
-            api_key=data.api_key,
-            base_url=data.base_url,
-            model_name=data.model_name,
-            enabled=data.enabled,
-            is_default=data.is_default,
-            config=data.config,
-            description=data.description,
-        )
-
-        # Audit log
-        await async_audit_logger.log(
-            action="create",
-            resource="cloud_provider",
-            resource_id=provider["id"],
-            user_id=user.username,
-            details={"name": data.name, "provider_type": data.provider_type},
-        )
-
-        return {"status": "ok", "provider": provider}
-    except Exception as e:
-        raise HTTPException(status_code=400, detail=str(e))
-
-
-@app.put("/admin/llm/providers/{provider_id}")
-async def admin_update_cloud_provider(
-    provider_id: str, data: CloudProviderUpdate, user: User = Depends(get_current_user)
-):
-    """Update cloud LLM provider"""
-    # Filter out None values
-    update_data = {k: v for k, v in data.model_dump().items() if v is not None}
-
-    provider = await async_cloud_provider_manager.update_provider(provider_id, **update_data)
-    if not provider:
-        raise HTTPException(status_code=404, detail="Provider not found")
-
-    # Audit log
-    await async_audit_logger.log(
-        action="update",
-        resource="cloud_provider",
-        resource_id=provider_id,
-        user_id=user.username,
-        details=update_data,
-    )
-
-    return {"status": "ok", "provider": provider}
-
-
-@app.delete("/admin/llm/providers/{provider_id}")
-async def admin_delete_cloud_provider(provider_id: str, user: User = Depends(get_current_user)):
-    """Delete cloud LLM provider"""
-    if not await async_cloud_provider_manager.delete_provider(provider_id):
-        raise HTTPException(status_code=404, detail="Provider not found")
-
-    # Audit log
-    await async_audit_logger.log(
-        action="delete",
-        resource="cloud_provider",
-        resource_id=provider_id,
-        user_id=user.username,
-    )
-
-    return {"status": "ok", "message": f"Provider {provider_id} deleted"}
-
-
-@app.post("/admin/llm/providers/{provider_id}/test")
-async def admin_test_cloud_provider(provider_id: str, user: User = Depends(get_current_user)):
-    """Test cloud provider connection"""
-    provider_config = await async_cloud_provider_manager.get_provider_with_key(provider_id)
-    if not provider_config:
-        raise HTTPException(status_code=404, detail="Provider not found")
-
-    if not provider_config.get("api_key"):
-        return {
-            "status": "error",
-            "available": False,
-            "message": "No API key configured",
-        }
-
-    try:
-        service = CloudLLMService(provider_config)
-        is_available = service.is_available()
-
-        if is_available:
-            # Quick test generation
-            test_response = service.generate_response("Скажи 'тест ок'", use_history=False)
-            return {
-                "status": "ok",
-                "available": True,
-                "test_response": test_response[:200] if test_response else "",
-            }
-        else:
-            return {
-                "status": "error",
-                "available": False,
-                "message": "Provider not responding",
-            }
-    except Exception as e:
-        return {
-            "status": "error",
-            "available": False,
-            "message": str(e),
-        }
-
-
-@app.post("/admin/llm/providers/{provider_id}/set-default")
-async def admin_set_default_cloud_provider(
-    provider_id: str, user: User = Depends(get_current_user)
-):
-    """Set cloud provider as default"""
-    if not await async_cloud_provider_manager.set_default(provider_id):
-        raise HTTPException(status_code=404, detail="Provider not found or disabled")
-
-    await async_audit_logger.log(
-        action="update",
-        resource="cloud_provider",
-        resource_id=provider_id,
-        user_id=user.username,
-        details={"is_default": True},
-    )
-
-    return {"status": "ok", "message": f"Provider {provider_id} set as default"}
-
-
-@app.get("/admin/llm/personas")
-async def admin_get_personas():
-    """Получить список доступных персон"""
-    if llm_service and hasattr(llm_service, "get_available_personas"):
-        return {"personas": llm_service.get_available_personas()}
-
-    # Fallback для Gemini LLM Service
-    from vllm_llm_service import SECRETARY_PERSONAS
-
-    return {
-        "personas": {
-            pid: {"name": p["name"], "full_name": p.get("full_name", p["name"])}
-            for pid, p in SECRETARY_PERSONAS.items()
-        }
-    }
-
-
-@app.get("/admin/llm/persona")
-async def admin_get_current_persona():
-    """Получить текущую персону"""
-    if llm_service:
-        persona_id = getattr(llm_service, "persona_id", "anna")
-        persona = getattr(llm_service, "persona", {})
-        return {
-            "id": persona_id,
-            "name": persona.get("name", "Unknown"),
-        }
-    return {"id": "none", "error": "LLM service not initialized"}
-
-
-@app.post("/admin/llm/persona")
-async def admin_set_persona(request: AdminPersonaRequest, user: User = Depends(get_current_user)):
-    """Установить персону"""
-    if llm_service and hasattr(llm_service, "set_persona"):
-        success = llm_service.set_persona(request.persona)
-        if success:
-            # Audit log
-            await async_audit_logger.log(
-                action="update",
-                resource="config",
-                resource_id="llm_persona",
-                user_id=user.username,
-                details={"persona": request.persona},
-            )
-            return {"status": "ok", "persona": request.persona}
-        raise HTTPException(status_code=400, detail=f"Persona not found: {request.persona}")
-    raise HTTPException(status_code=503, detail="LLM service does not support personas")
-
-
-@app.get("/admin/llm/params")
-async def admin_get_llm_params():
-    """Получить параметры генерации LLM"""
-    if llm_service and hasattr(llm_service, "runtime_params"):
-        return {"params": llm_service.runtime_params}
-
-    # Возвращаем значения по умолчанию
-    return {
-        "params": {"temperature": 0.7, "max_tokens": 512, "top_p": 0.9, "repetition_penalty": 1.1}
-    }
-
-
-@app.post("/admin/llm/params")
-async def admin_set_llm_params(request: AdminLLMParamsRequest):
-    """Установить параметры генерации LLM"""
-    if llm_service and hasattr(llm_service, "set_params"):
-        params = {k: v for k, v in request.dict().items() if v is not None}
-        llm_service.set_params(**params)
-        return {"status": "ok", "params": llm_service.runtime_params}
-
-    # Для vLLM сервиса без set_params - сохраняем в атрибуте
-    if llm_service:
-        if not hasattr(llm_service, "runtime_params"):
-            llm_service.runtime_params = {}
-        params = {k: v for k, v in request.dict().items() if v is not None}
-        llm_service.runtime_params.update(params)
-        return {"status": "ok", "params": llm_service.runtime_params}
-
-    raise HTTPException(status_code=503, detail="LLM service not initialized")
-
-
-@app.get("/admin/llm/prompt/{persona}")
-async def admin_get_persona_prompt(persona: str):
-    """Получить системный промпт для персоны"""
-    try:
-        from vllm_llm_service import SECRETARY_PERSONAS
-
-        if persona in SECRETARY_PERSONAS:
-            return {"persona": persona, "prompt": SECRETARY_PERSONAS[persona]["prompt"]}
-        raise HTTPException(status_code=404, detail=f"Persona not found: {persona}")
-    except ImportError:
-        raise HTTPException(status_code=503, detail="vLLM service not available")
-
-
-@app.post("/admin/llm/prompt/{persona}")
-async def admin_set_persona_prompt(persona: str, request: AdminLLMPromptRequest):
-    """Установить системный промпт для персоны"""
-    try:
-        from vllm_llm_service import SECRETARY_PERSONAS
-
-        if persona not in SECRETARY_PERSONAS:
-            raise HTTPException(status_code=404, detail=f"Persona not found: {persona}")
-
-        # Обновляем промпт
-        SECRETARY_PERSONAS[persona]["prompt"] = request.prompt
-
-        # Если это текущая персона - обновляем в сервисе
-        if llm_service and hasattr(llm_service, "persona_id") and llm_service.persona_id == persona:
-            llm_service.system_prompt = request.prompt
-
-        return {"status": "ok", "persona": persona}
-    except ImportError:
-        raise HTTPException(status_code=503, detail="vLLM service not available")
-
-
-@app.post("/admin/llm/prompt/{persona}/reset")
-async def admin_reset_persona_prompt(persona: str):
-    """Сбросить системный промпт персоны на значение по умолчанию"""
-    # TODO: Реализовать хранение оригинальных промптов
-    raise HTTPException(status_code=501, detail="Not implemented yet")
-
-
-# ============== Monitoring Endpoints ==============
-
-
-@app.get("/admin/monitor/gpu")
-async def admin_get_gpu_stats():
-    """Получить статистику GPU"""
-    import torch
-
-    if not torch.cuda.is_available():
-        return {"available": False, "gpus": []}
-
-    gpus = []
-    for i in range(torch.cuda.device_count()):
-        try:
-            name = torch.cuda.get_device_name(i)
-            props = torch.cuda.get_device_properties(i)
-            total_memory = props.total_memory / (1024**3)
-            allocated = torch.cuda.memory_allocated(i) / (1024**3)
-            reserved = torch.cuda.memory_reserved(i) / (1024**3)
-
-            # Пытаемся получить утилизацию через nvidia-smi
-            utilization = None
-            temperature = None
-            try:
-                import subprocess
-
-                result = subprocess.run(
-                    [
-                        "nvidia-smi",
-                        f"--id={i}",
-                        "--query-gpu=utilization.gpu,temperature.gpu",
-                        "--format=csv,noheader,nounits",
-                    ],
-                    capture_output=True,
-                    text=True,
-                    timeout=5,
-                )
-                if result.returncode == 0:
-                    parts = result.stdout.strip().split(",")
-                    if len(parts) >= 2:
-                        utilization = int(parts[0].strip())
-                        temperature = int(parts[1].strip())
-            except Exception:
-                pass
-
-            gpus.append(
-                {
-                    "id": i,
-                    "name": name,
-                    "total_memory_gb": round(total_memory, 2),
-                    "allocated_gb": round(allocated, 2),
-                    "reserved_gb": round(reserved, 2),
-                    "free_gb": round(total_memory - reserved, 2),
-                    "utilization_percent": utilization,
-                    "temperature_c": temperature,
-                    "compute_capability": f"{props.major}.{props.minor}",
-                }
-            )
-        except Exception as e:
-            logger.warning(f"Error getting GPU {i} stats: {e}")
-
-    return {"available": True, "gpus": gpus}
-
-
-@app.get("/admin/monitor/health")
-async def admin_get_health():
-    """Расширенная проверка здоровья всех компонентов"""
-    manager = get_service_manager()
-    health = {"timestamp": datetime.now().isoformat(), "overall": "healthy", "components": {}}
-
-    # Orchestrator
-    health["components"]["orchestrator"] = {"status": "healthy", "uptime": "running"}
-
-    # LLM
-    if llm_service:
-        try:
-            if isinstance(llm_service, CloudLLMService):
-                health["components"]["llm"] = {
-                    "status": "healthy",
-                    "backend": f"cloud:{getattr(llm_service, 'provider_id', 'unknown')}",
-                }
-            elif hasattr(llm_service, "is_available") and llm_service.is_available():
-                health["components"]["llm"] = {"status": "healthy", "backend": "vllm"}
-            else:
-                health["components"]["llm"] = {"status": "healthy", "backend": "unknown"}
-        except Exception as e:
-            health["components"]["llm"] = {"status": "unhealthy", "error": str(e)}
-            health["overall"] = "degraded"
-    else:
-        health["components"]["llm"] = {"status": "unavailable"}
-        health["overall"] = "degraded"
-
-    # TTS
-    if anna_voice_service or voice_service:
-        health["components"]["tts_xtts"] = {"status": "healthy"}
-    else:
-        health["components"]["tts_xtts"] = {"status": "unavailable"}
-
-    if piper_service:
-        health["components"]["tts_piper"] = {"status": "healthy"}
-    else:
-        health["components"]["tts_piper"] = {"status": "unavailable"}
-
-    # vLLM external process
-    vllm_status = manager.get_service_status("vllm")
-    if vllm_status["is_running"]:
-        health["components"]["vllm_process"] = {"status": "healthy", "pid": vllm_status["pid"]}
-    else:
-        health["components"]["vllm_process"] = {"status": "stopped"}
-
-    return health
-
-
-@app.get("/admin/monitor/metrics")
-async def admin_get_metrics():
-    """Получить метрики системы"""
-    import psutil
-
-    metrics = {
-        "timestamp": datetime.now().isoformat(),
-        "system": {
-            "cpu_percent": psutil.cpu_percent(interval=0.1),
-            "memory_percent": psutil.virtual_memory().percent,
-            "memory_used_gb": round(psutil.virtual_memory().used / (1024**3), 2),
-            "memory_total_gb": round(psutil.virtual_memory().total / (1024**3), 2),
-        },
-        "streaming_tts": streaming_tts_manager.get_stats() if streaming_tts_manager else None,
-    }
-
-    # LLM метрики
-    if llm_service:
-        metrics["llm"] = {
-            "history_length": len(getattr(llm_service, "conversation_history", [])),
-            "faq_count": len(getattr(llm_service, "faq", {})),
-        }
-
-    return metrics
-
-
-@app.get("/admin/monitor/errors")
-async def admin_get_errors():
-    """Получить последние ошибки"""
-    manager = get_service_manager()
-    return {"errors": manager.last_errors, "timestamp": datetime.now().isoformat()}
-
-
-@app.post("/admin/monitor/metrics/reset")
-async def admin_reset_metrics():
-    """Сбросить метрики"""
-    # Очищаем кэш TTS
-    if streaming_tts_manager:
-        with streaming_tts_manager._cache_lock:
-            streaming_tts_manager._cache.clear()
-
-    return {"status": "ok", "message": "Metrics reset"}
-
-
-@app.get("/admin/monitor/system")
-async def admin_get_system_status():
-    """Полная информация о системе: GPU, CPU, RAM, диски, Docker, сеть"""
-    monitor = get_system_monitor()
-    return monitor.get_full_status()
-
-
-# ============== Model Management API ==============
-
-
-@app.get("/admin/models/list")
-async def admin_list_models():
-    """Список всех локальных моделей"""
-    manager = get_model_manager()
-    return {"models": manager.get_cached_models()}
-
-
-@app.post("/admin/models/scan")
-async def admin_scan_models(request: Request):
-    """Запустить сканирование моделей"""
-    data = await request.json() if request.headers.get("content-type") == "application/json" else {}
-    include_system = data.get("include_system", False)
-
-    manager = get_model_manager()
-    if manager.scan_all_models(include_system=include_system):
-        return {"status": "ok", "message": "Scan started"}
-    else:
-        return {"status": "error", "message": "Scan already in progress"}
-
-
-@app.post("/admin/models/scan/cancel")
-async def admin_cancel_scan():
-    """Отменить сканирование"""
-    manager = get_model_manager()
-    manager.cancel_scan()
-    return {"status": "ok", "message": "Scan cancelled"}
-
-
-@app.get("/admin/models/scan/status")
-async def admin_scan_status():
-    """Статус сканирования"""
-    manager = get_model_manager()
-    return {"status": manager.get_scan_progress()}
-
-
-@app.post("/admin/models/download")
-async def admin_download_model(request: Request):
-    """Скачать модель с HuggingFace"""
-    data = await request.json()
-    repo_id = data.get("repo_id")
-    revision = data.get("revision", "main")
-
-    if not repo_id:
-        raise HTTPException(status_code=400, detail="repo_id required")
-
-    manager = get_model_manager()
-    if manager.download_model(repo_id, revision):
-        return {"status": "ok", "message": f"Download started: {repo_id}"}
-    else:
-        return {"status": "error", "message": "Download already in progress"}
-
-
-@app.post("/admin/models/download/cancel")
-async def admin_cancel_download():
-    """Отменить загрузку"""
-    manager = get_model_manager()
-    manager.cancel_download()
-    return {"status": "ok", "message": "Download cancelled"}
-
-
-@app.get("/admin/models/download/status")
-async def admin_download_status():
-    """Статус загрузки"""
-    manager = get_model_manager()
-    return {"status": manager.get_download_progress()}
-
-
-@app.delete("/admin/models/delete")
-async def admin_delete_model(path: str):
-    """Удалить модель"""
-    manager = get_model_manager()
-    result = manager.delete_model(path)
-    if result["status"] == "ok":
-        return result
-    else:
-        raise HTTPException(status_code=400, detail=result.get("error", "Delete failed"))
-
-
-@app.get("/admin/models/search")
-async def admin_search_huggingface(query: str, limit: int = 20):
-    """Поиск моделей на HuggingFace"""
-    manager = get_model_manager()
-    results = manager.search_huggingface(query, limit)
-    return {"results": results}
-
-
-@app.get("/admin/models/details/{repo_id:path}")
-async def admin_get_model_details(repo_id: str):
-    """Получить детали модели с HuggingFace"""
-    manager = get_model_manager()
-    details = manager.get_model_details(repo_id)
-    if details:
-        return {"details": details}
-    else:
-        raise HTTPException(status_code=404, detail="Model not found")
-
-
-# Widget public endpoints moved to modules/channels/widget/router_public.py
-from modules.channels.widget.router_public import router as widget_public_router  # noqa: E402
-
-
-app.include_router(widget_public_router)
-
-
+# Remaining admin endpoints extracted in Phase 4.5
 # ============== Static Files for Vue Admin ==============
 
 DEV_MODE = os.getenv("DEV_MODE", "").lower() in ("1", "true", "yes")
